Update the Presentation script
This unifies the enter/exit modes and provides more feedback. It will
toss up a dialog, which you can just tap enter at to accept the default of
switching modes, or you can re-enable/disable modes in case something
went wrong. It currently also confirms with a dialog, because growl
breaks so often. If growl is working, it should return immediately, if
not you know to get your shotgun out.

This also switches to using caffeinate instead of pmset, so no more
passwords in your scripts. Yay!